VOK_02
On this version, the interfaces were reviewed and fixed; the tests for the shuffles were updated; the class "BancoDePalavras" was changed to insert a method getBancoDePalavras that allow to work on a copy of the original words array.
The code of "MecanicaDoJogo" was refactored to reuse part of the code, then a private method was created.
There are more things to do such as: clean the code, add more comments, read the word from a file (optional, but interesting) and to update the UML diagram.