VOK_03
On this version the game is reading the words from a file (in this case "frutas.txt"), but the old script was held there (as old).
Also, it was verified that the list of words showed at the end of the game was including words from the previous matches. It was included a method to clear the lists every time that the game is started.